+++ brue/js_reserves.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+class _This(object):
+    """
+    javascript this reserved words
+    """
+
+this = _This()